+++ nlp4all/controllers/UserController.py
@@ -1,6 +1,7 @@
 """User / Login related controller."""
 
 import os
+from random import randint
 import secrets
 from PIL import Image
 from flask_bcrypt import check_password_hash, generate_password_hash
@@ -9,9 +10,11 @@ from flask import flash, redirect, render_template, request, url_for, Blueprint
 from flask_login import current_user, login_user, logout_user
 
 from nlp4all.models.database import db_session
-from nlp4all.models import Organization, User
+from nlp4all.models import BayesianAnalysis, Organization, User
 
-from nlp4all.forms.user import LoginForm, UpdateAccountForm, RegistrationForm
+from nlp4all.forms.user import LoginForm, UpdateAccountForm, RegistrationForm, RequestResetForm, ResetPasswordForm, IMCRegistrationForm
+
+from flask_mail import Message
 
 
 class UserController:
@@ -101,14 +104,86 @@ class UserController:
         return render_template("register.html", title="Register", form=form)
 
     @classmethod
-    def reset_request():
+    def reset_request(cls):
         """Reset password request page"""
         if current_user.is_authenticated:
             return redirect(url_for("home"))
         form = RequestResetForm()
         if form.validate_on_submit():
             user = User.query.filter_by(email=form.email.data).first()
-            send_reset_email(user)
+            cls.send_reset_email(user)
             flash("An email has been sent with instructions to reset your password.", "info")
             return redirect(url_for("login"))
         return render_template("reset_request.html", title="Reset Password", form=form)
+
+    @classmethod
+    def reset_token(cls, token):
+        """Reset password token page"""
+        if current_user.is_authenticated:
+            return redirect(url_for("home"))
+        user = User.verify_reset_token(token, cls.blueprint.secret_key)
+        if user in ('Expired', 'Invalid'):
+            flash(f"That is an {user} token", "warning")
+            return redirect(url_for("reset_request"))
+        form = ResetPasswordForm()
+        if form.validate_on_submit():
+            hashed_password = generate_password_hash(
+                form.password.data).decode("utf-8")
+            user.password = hashed_password
+            db_session.commit()
+            flash("Your password has been updated! You are now able to log in", "success")
+            return redirect(url_for("login"))
+        return render_template("reset_token.html", title="Reset Password", form=form)
+
+    @classmethod
+    def register_imc(cls):
+        """IMC registration page"""
+        if current_user.is_authenticated:
+            return redirect(url_for("home"))
+        form = IMCRegistrationForm()
+        if form.validate_on_submit():
+            fake_id = randint(0, 99999999999)
+            fake_email = str(fake_id) + "@arthurhjorth.com"
+            fake_password = str(fake_id)
+            hashed_password = generate_password_hash(
+                fake_password).decode("utf-8")
+            imc_org = Organization.query.filter_by(name="ATU").all()
+            a_project = imc_org[0].projects[0] #error when no project. out of range TODO
+            the_name = form.username.data
+            if any(User.query.filter_by(username=the_name)):
+                the_name = the_name + str(fake_id)
+            user = User(
+                username=the_name, email=fake_email, password=hashed_password, organizations=imc_org
+            )
+            db_session.add(user)
+            db_session.commit()
+            login_user(user)
+            userid = current_user.id
+            name = current_user.username + "'s personal analysis"
+            bayes_analysis = BayesianAnalysis(
+                user=userid,
+                name=name,
+                project=a_project.id,
+                data={"counts": 0, "words": {}},
+                tweets=[],
+                annotation_tags={},
+                annotate=1,
+            )
+            db.session.add(bayes_analysis)
+            db.session.commit()
+            return redirect(url_for("home"))
+        return render_template("register_imc.html", form=form)
+
+    @classmethod
+    def send_reset_email(cls, user):
+        """Send reset email"""
+        token = user.get_reset_token()
+        msg = Message("Password Reset Request",
+                    sender="noreply@demo.com", recipients=[user.email])
+        msg.body = f"""To reset your password, visit the following link:
+{url_for('reset_token', token=token, _external=True)}
+
+If you did not make this request then simply ignore this email and no changes will be made.
+"""
+        # @TODO implement mail server
+        # mail.send(msg)
