+++ nlp4all/controllers/__init__.py
@@ -2,5 +2,7 @@
 
 # Add classes to top level of module
 from .UserController import UserController
-
-
+from .AdminController import AdminController
+from .SiteController import SiteController
+from .ProjectController import ProjectController
+from .AnalysesController import AnalysesController
