+++ nlp4all/forms/user.py
@@ -1,3 +1,5 @@
+"""User related forms"""
+
 from flask_wtf import FlaskForm
 from flask_wtf.file import FileField, FileAllowed
 from flask_login import current_user
@@ -83,3 +85,39 @@ class RegistrationForm(FlaskForm):
         user = User.query.filter_by(email=email.data).first()
         if user:
             raise ValidationError("That email is taken. Please choose a different one.")
+
+
+class IMCRegistrationForm(FlaskForm):
+    """
+    IMC Registration form.
+    """
+    username = StringField("Username", validators=[DataRequired(), Length(min=2, max=20)])
+    submit = SubmitField("Sign Up")
+
+
+
+class RequestResetForm(FlaskForm):
+    """
+    Request reset form.
+    """
+    email = StringField("Email", validators=[DataRequired(), Email()])
+    submit = SubmitField("Request Password Reset")
+
+    def validate_email(self, email):
+        """
+        Validate email.
+        """
+        user = User.query.filter_by(email=email.data).first()
+        if user is None:
+            raise ValidationError("There is no account with that email. You must register first.")
+
+
+class ResetPasswordForm(FlaskForm):
+    """
+    Reset password form.
+    """
+    password = PasswordField("Password", validators=[DataRequired()])
+    confirm_password = PasswordField(
+        "Confirm Password", validators=[DataRequired(), EqualTo("password")]
+    )
+    submit = SubmitField("Reset Password")
