+++ pyproject.toml
@@ -21,4 +21,4 @@ disable = [
 ]
 
 [tool.mypy]
-plugins = sqlmypy
+plugins = "sqlmypy"
